Separate paging initialization from kernel.
The paging initialization code is relatively small and is unlikely to
ever exceed one 4MB page. However, the kernel code/data might eventually
grow past the page boundary, which would require the
`boot_page_directory` to be manually updated to map some more pages.
Manually writing page tables is obviously undesirable, so the init logic
which sets up the initial page tables needed to bootstrap the kernel
needs to be separated from the main part of the kernel. This is the
first part of the extraction (the second part will update the init code
to create the "real" page tables before jumping to the kernel).

This commit also makes it possible to debug the kernel entrypoint (this
was previously difficult because *all* the code was mapped into the
higher-half).