Set up final page directory before jumping to higher half.
This moves the kernel page directory set-up to `init_paging`, which is
called in the bootstrapping phase before the kernel begins executing in
the higher half.

Also, instead of mapping the entire 1GB address space, the kernel now
only maps however much is needed for the text/data sections and for the
heap.